shill: don't process proxy packet after read error
IOInputHandler::OnFileCanReadWithoutBlocking calls the error handler
on a read error, but it also calls the input handler afterwards.
We should really change the behavior of the IOInputHandler,
but I don't know if some callers rely on the current behavior, so
I am fixing this only for the HTTP proxy case.  Also clean up the
confusion between ssize_t, size_t, and int (everything should be
ssize_t here).

BUG=crosbug.com/p/52689
TEST=none

+++ shill/net/io_input_handler.cc
@@ -51,7 +51,7 @@ void IOInputHandler::OnFileCanReadWithoutBlocking(int fd) {
   CHECK_EQ(fd_, fd);
 
   unsigned char buf[IOHandler::kDataBufferSize];
-  int len = read(fd, buf, sizeof(buf));
+  ssize_t len = read(fd, buf, sizeof(buf));
   if (len < 0) {
     std::string condition = base::StringPrintf(
         "File read error: %d", errno);
